requirements.txt
nicegui>=1.4.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
torch>=2.0.0
transformers>=4.30.0
scipy>=1.7.0
segmentation-models-pytorch>=0.3.0
opencv-python>=4.8.0
albumentations>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
